refactor(main): DiffAppState accessor を state module に分離
DIFF_APP_STATE thread-local と with_app_state を state.rs に移し、main.rs を composition root へ近づける。

Refs: #224
